\name{gls.simulate}
\alias{gls.simulate}

\title{ Simulation data of the Group lasso model}

\description{ 
Generating simulation data object by the Group lasso model (GLS model). 
} 

\usage{ 
gls.simulate(file.phe.out, 
      file.snp.out, 
      simu_grp=1, 
      simu_n=500,
      simu_p=1000, 
      simu_snp_rho = 0.1, 
      simu_snp_missing= 0.002, 
      simu_rho = 0.4, 
      simu_sigma2 = 16,
      simu_mu = c(13.395, -3.08, 1.875, -3.195), 
      simu_cov_range = c( -1, 1 ),
      simu_cov_effect= array(c(0,0,0,0), dim=c(1,4)), 
      simu_add_pos   = c( 1,2,3 ), 
      simu_add_effect= array(c( 1.04, 0.885, -2.055, 0.545, 1.17, 
      -0.20, 0.74, -4.715,1.40, -2.25, 1.00,  0.00), dim=c(3,4)),
      simu_dom_pos = c( 3,4,5 ), 
      simu_dom_effect= array(c( 1.49, -2.135, 4.82, 1.425, 1.045, 
      1.320, 1.905, 1.535, 1.265, -1.225, 2.710, -1.96),dim=c(3,4)),
      simu_z_range = c(20,80), simu_z_count = c( 5, 12 ), 
      plink.format = FALSE,
      debug = FALSE )}

\arguments{
\item{file.phe.out}{ String, the name of the output phenotypic data file. }
\item{file.snp.out}{ String, the name of the output genotypic data file. }
\item{simu_grp}{ Numeric, the number of groups that snps are divided into for parallel computation. \cr Default: 1.}
\item{simu_n}{ Integer, sample size. \cr Default: 500.}
\item{simu_p}{ Integer, number of SNP. \cr Default: 1000.}
\item{simu_snp_rho}{ Float, correlation coefficient between two adjacent SNPs. \cr Default: 0.1.}
\item{simu_snp_missing}{Float, ratio of missing SNPs.}
\item{simu_rho}{ Float, correlation coefficient between two adjacent time points. \cr Default: 0.4.}
\item{simu_sigma2}{ Float, residual error. \cr Default: 16.}
\item{simu_mu}{ Vector, Legendre coefficients of overall mean effect. \cr Default: c(13.395,-3.08, 1.875, -3.195).}
\item{simu_cov_range}{ Vector, range of covariates. \cr Default: c(-1,1). }
\item{simu_cov_effect}{Vector, Legendre coefficients of for covariates. \cr Default: array(c(0,0,0,0), dim=c(1,4).} for 1 covariate.
\item{simu_add_pos}{ Vector, positions of the significant SNPs with additive effects. \cr Default: c(1,2,3 ).}
\item{simu_add_effect}{ Matrix, Legendre coefficients for additive effects of significant SNPs. 
    \cr Default: array(c( 1.04, 0.885, -2.055, 0.545, 1.17, -0.20, 0.74, -4.715, 1.40, -2.25, 1.00,  0.00), dim=c(3,4)). }
\item{simu_dom_pos}{ Vector, positions of the significant SNPs with dominant effects. Default: c(3,4,5). }
\item{simu_dom_effect}{ Matrix, Legendre coefficients for dominant effects of significant SNPs. 
    \cr Default: array(c( 1.49, -2.135, 4.82, 1.425, 1.045, 1.320, 1.905, 1.535,1.265, -1.225, 2.710, -1.96), dim=c(3,4)). }
\item{simu_z_range}{ Vector, range of measure times. \cr Default: c(20, 80).}
\item{simu_z_count}{ Integer, the range of measure count. \cr Default: c(5, 12 ).}
\item{plink.format}{ Logical, if TRUE, save the SNP data as PLINK format. \cr Default: FALSE. }
\item{debug}{ Logical, if TRUE, run the command in debug model which inputs more messages for debugging. \cr Default: FALSE. }
}

\value{
Returns two simulation data files: the phenotypic data file and genotypic data file. 
}

\details{
The simulation in this package uses the pre-defined parameters to create a data object containing 
longitudinal phenotypic data and genotypic data. \cr

The genotypic data file is coded 2, 1, 0, -1 to indicate genotypes QQ, Qq, qq and missing data respectively.\cr

It is assumed that the trait is controlled by some SNPs jointly, and the positions of these SNPs will be generated 
by the parameters \code{simu_a_pos} and \code{simu_d_pos}. 
The genetic effects of overall mean and causal SNPs will be generated by the parameters \code{simu_mu}, 
\code{simu_a_effect} and \code{simu_d_effect}.\cr

Given phenotypic data and genotype information, genetic effects of each SNPs could be estimated. 
However in GWAS a number of covariates either discrete or continuous may be measured for each subject. 
In the simulation function, covariate effects are generated by parameter \code{simu_covar_effect}.\cr

In particular, since measurements within each subject are possibly correlated with one another, 
AR(1) model is employed to approximate the residual covariance matrix, and assume AR(1) with \code{simu_rho} and \code{simu_sigma2}. \cr
}

\author{Zhong Wang and Nating Wang }

\references{
 Li, J., Wang, Z., Li, R., & Wu, R. (2015). Bayesian group Lasso for nonparametric varying-coefficient models with application to functional genome-wide association studies. The Annals of Applied Statistics, 9(2), 640-664.\cr
}

\examples{

gls.simulate( "gls.test.simple.phe", "gls.test.simple.snp", plink.format=TRUE );


}

\keyword{functions}
\keyword{GLS model}
\keyword{simulate}
